requirements.txt
alembic==1.12.0
blinker==1.6.3
blueprint==3.4.2
Bootstrap-Flask==2.3.2
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
dominate==2.8.0
elastic-transport==8.10.0
elasticsearch==8.11.0
email-validator==2.1.0.post1
Flask==2.2.5
Flask-Admin==1.6.1
flask-bootstrap5==0.1.dev1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.1.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.1.0
